fix code after codereview/change images bg/add in data object with svg icons
