+++ components/HallFameList/HallFameList.module.css
@@ -1,23 +1,27 @@
 .hallFameListContainer {
-  background-image: url('/images/hallfame/bghallfame.webp');
+  background-image: image-set(
+    url('/images/hallfame/bghallfamemob.webp') 1x,
+    url('/images/hallfame/bghallfamemob@2x.webp') 2x
+  );
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
 }
 
-@media (min-width: 768px) {
+@media screen and (min-width: 768px) {
   .hallFameListContainer {
-    background-image: url('/images/hallfame/bghallfametabl.webp');
-  }
-  @media (min-resolution: 192dpi), (min-width: 2dppx) {
-    .hallFameListContainer {
-      background-image: url('/images/hallfame/bghallfametabl@2x.webp');
-    }
+    background-image: image-set(
+      url('/images/hallfame/bghallfametab.webp') 1x,
+      url('/images/hallfame/bghallfametab@2x.webp') 2x
+    );
   }
 }
 
-@media (min-resolution: 192dpi), (min-width: 2dppx) {
+@media screen and (min-width: 1280px) {
   .hallFameListContainer {
-    background-image: url('/images/hallfame/bghallfame@2x.webp');
+    background-image: image-set(
+      url('/images/hallfame/bghallfamedesc.webp') 1x,
+      url('/images/hallfame/bghallfamedesc@2x.webp') 2x
+    );
   }
 }
