+++ components/HallFameList/HallFameList.tsx
@@ -1,16 +1,17 @@
 'use client';
 import React from 'react';
 import { useState, useEffect } from 'react';
+
 import HallFameCard from '@/components/HallFameCard/HallFameCard';
 import HallFameCustomCard from '@/components/HallFameCustomCard/HallFameCustomCard';
-import hallFameHeading from '@/data/hallFameHeading.json';
-
-import { getHallFameCards } from '@/utils/getHallFameCards';
-import customCardData from '@/data/customCardData.json';
-import s from '@/components/HallFameList/HallFameList.module.css';
-import { HallFameCardProps, HallFameCustomCardProps } from '@/types/index';
 import { Heading } from '@/components/Heading/Heading';
 import { Section } from '@/components/Section/Section';
+
+import { getHallFameCards } from '@/utils/getHallFameCards';
+import s from '@/components/HallFameList/HallFameList.module.css';
+import { HallFameCardProps, HallFameCustomCardProps } from '@/types/index';
+import hallFameHeading from '@/data/hallFameHeading.json';
+import customCardData from '@/data/customCardData.json';
 
 const isHallFameCardProps = (
   data: HallFameCardProps | HallFameCustomCardProps | undefined,
@@ -39,7 +40,7 @@ const HallFameList: React.FC = () => {
         const data = await getHallFameCards();
         setHallFameData(data);
       } catch (error) {
-        console.error('Error fetching data in component:', error);
+        console.error('Error fetching data in component:', error); //TO-DO обробити помилку коли зявиться картинка-заглушка
       }
     };
