+++ data/hallFameCardIconMap.ts
@@ -1,0 +1,13 @@
+import IconFirst from '@/public/images/hallfame/iconcupfirst.svg';
+import IconSecond from '@/public/images/hallfame/iconcuptwo.svg';
+import IconThird from '@/public/images/hallfame/iconcupthree.svg';
+import IconStar from '@/public/images/hallfame/iconcupstar.svg';
+
+const iconsMap = {
+  first: IconFirst,
+  second: IconSecond,
+  third: IconThird,
+  star: IconStar,
+};
+
+export default iconsMap;
